requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiolimiter==1.2.1
aiosignal==1.3.2
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.3
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.4.4
Pillow==11.2.1
propcache==0.3.2
python-dotenv==1.1.0
python-telegram-bot==22.1
pyrogram==2.0.106
tgcrypto==1.2.5
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.20.0